+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## x.x.x x-x-x
+### PaymentSheet
+* [Added] Embedded Payment Element is now generally available. Learn more by visiting the [docs](https://docs.stripe.com/payments/mobile/embedded).
+
 ## 24.12.1 2025-04-30
 ### Identity
 * [Fixed] Fixed an issue where identity verification may fail due to stale model cache.